Add files via upload
Ten minute radio light curves of 4U 1728-34. The MJD provided is the mid-point of the 10-minute long time bin. S5.5 is the flux density at 5.5GHz in mJy, S5.5_err is the 1-sigma uncertainty in mJy. S9 is the flux density at 9GHz in mJy, S9_err is the 1-sigma uncertainty in mJy.