Add files via upload
Averaged 1-minute light curves of 4U 1728-34, showing the radio flux density averaged over all bursts. The time provided is the mid-point of the 1-minute long time bin, where time=0 is the start of the bursts. S5.5 is the flux density at 5.5GHz in mJy, S5.5_err is the 1-sigma uncertainty in mJy. S9 is the flux density at 9GHz in mJy, S9_err is the 1-sigma uncertainty in mJy.