Add files via upload
Individual 4-min lightcurve data for single bursts from 4U 1728-34.  The time provided is the mid-point of the 4-minute long time bin, where time=0 is the start of the bursts. S5.5 is the flux density at 5.5GHz in mJy, S5.5_err is the 1-sigma uncertainty in mJy. S9 is the flux density at 9GHz in mJy, S9_err is the 1-sigma uncertainty in mJy, burst gives the burst number.